+++ src/Morphant.Generator/ConstructionSurface/ConstructionPlan/ConstructionPlanEmitter.cs
@@ -32,13 +32,10 @@ internal static class ConstructionPlanEmitter
         var typeReference = BuildTypeReference(
             Identifier(model.ConstructorParametersTypeName),
             model.TypeParameters);
-        var destinationCref = XmlAttribute(
-            model.DestinationCref);
-
         WriteSummary(
             writer,
             "Maps constructor arguments for " +
-            $"<see cref=\"{destinationCref}\"/>.");
+            DestinationDocumentation(model) + ".");
         WriteTypeDeclaration(
             writer,
             $"internal sealed class {typeReference}",
@@ -204,11 +201,17 @@ internal static class ConstructionPlanEmitter
         CodeWriter writer,
         ConstructionPlanModel model)
     {
-        var cref = XmlAttribute(model.DestinationCref);
         WriteSummary(
             writer,
             "Defines construction of " +
-            $"<see cref=\"{cref}\"/>.");
+            DestinationDocumentation(model) + ".");
+    }
+
+    private static string DestinationDocumentation(ConstructionPlanModel model)
+    {
+        return model.DestinationDisplayName is { } display
+            ? $"<c>{XmlText(display)}</c>"
+            : $"<see cref=\"{XmlAttribute(model.DestinationCref)}\"/>";
     }
 
     private static void WriteDestinationConstructorDocumentation(